add standalone page to sandcastle v2
